рефакторинг main и popup
